Rename RucSounding tests to RucSoundingTests. Move value/interpolation tests into SoundingTests.
